# internal/handler/http/model/model_flat.go
/*
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package model

type Flat struct {
	Id int `json:"id"`

	HouseId int `json:"house_id"`

	Price int32 `json:"price"`

	Rooms int32 `json:"rooms"`

	Status Status `json:"status"`
}