# internal/handler/http/model/model_house.go
/*
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package model

import (
	"time"
)

type House struct {
	Id int `json:"id,omitempty"`

	Address string `json:"name" binding:"required"`

	Year int32 `json:"year" binding:"required"`

	Developer string `json:"developer,omitempty"`

	CreatedAt time.Time `json:"created_at,omitempty"`

	UpdateAt time.Time `json:"update_at,omitempty"`
}