# internal/handler/http/model/model_user_type.go
/*
 * Тестовое задание для отбора на Backend Bootcamp
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package model

// UserType : Тип пользователя
type UserType string

// List of UserType
const (
	CLIENT    UserType = "client"
	MODERATOR UserType = "moderator"
)